Added Target Grid Point Selection
The function selectTargetLocation() was added to combat tracker which
will open up selection of a grid unit.  This will be used for casting
spells.

Currently activates when using the Special action in battle, currently
no way to turn it off and it doesn't actually do anything other than
show your selection.