Folder cleanup and Animation
Moved some srcipts into separate folders to 1. see if this broke
anything in the game and 2. managing things.

changed the level indicator on top of enemies to only appear on the
world map. This was done by moving the control functions for the
indicator to thebasic enemy animations class from enemy behaviour.