More HEX fixes
Changes when we select a location to do something around.
Instead of creating vectors and checking if we move that far, what tile
are we on for moving along grid (like a fool), I will now keep the
"targetIntLocation" as coordinates on the map instead of being dumb.

Added some map routines for dealing with these coordinates instead of
the Vector3.