getPath has been made HEX!!
public List<int[]> getPath(int[] pos1, int[]pos2, int maxDist)  has had
a lot of work.

Now it will check around objects and generate the path of neighbours to
get between one point and another.

Need to find a way to move through the path of waypoints instead of A ->
B.  I also need to reduce MP correctly but that is a job for resetd
Hoog.