delayed steps in directions for creating the area where sight is blocked
more work on the Field of View generator.  I have added the ability to
delay "moves of darkness" when finding out which cells are blocked.  ie
cells directly away would have no delay, while the direction that is 50
degrees away will have a longer delay.  They still just affect people on
the direct line but this is getting closer.

I also added an indicator in the console for when yuou centrally senect
a location.  This will let me look at what the cube location of a spot
is by casting over it.