Out of range update
game now checks the tiles for whether or not a position is in range.
the tiles are set to be in range or not based on either field of view,
A* or unrestricted range.
Also added to the go Back function to make everything out of range if
you hit back to avoid exploits.